Revert "[SandboxIR][NFC] Move Region from SandboxVectorizer to SandboxIR. (#1…"
This reverts commit 3d9ed92630fb2a3282ba4a49b43d6eccca2f3509.
