+++ llvm/unittests/Transforms/Vectorize/SandboxVectorizer/CMakeLists.txt
@@ -11,4 +11,5 @@ add_llvm_unittest(SandboxVectorizerTests
   DependencyGraphTest.cpp
   InstrIntervalTest.cpp
   LegalityTest.cpp
+  RegionTest.cpp
   )
